+++ Essa/GUI/Widgets/Textfield.hpp
@@ -18,9 +18,9 @@ public:
     CREATE_VALUE(Util::UString, content, "")
     CREATE_VALUE(size_t, font_size, theme().label_font_size)
     CREATE_VALUE(Align, alignment, Align::CenterLeft)
-    CREATE_VALUE(float, padding, 5)
+    CREATE_VALUE(int, padding, 5)
 
-    Util::Rectf text_rect() const;
+    Util::Recti text_rect() const;
 
 private:
     virtual LengthVector initial_size() const override;
