+++ examples/essa-engine/obj-loader-resource-manager.cpp
@@ -12,15 +12,21 @@ class WorldView : public GUI::WorldView {
 public:
     explicit WorldView()
         : GUI::WorldView() {
-        // FIXME: The "background" would now cover cubes because it is rendered by
-        //        Builder AFTER draw() is called. Make background rendered immediately
-        //        for WorldView so that is is actually *below* the world render.
+        // FIXME: The "background" would now cover cubes because it is rendered
+        // by Builder AFTER draw() is called. Make background rendered
+        // immediately for WorldView so that is is actually *below* the
+        // world render.
         // set_background_color(Util::Colors::LightBlue);
     }
 
 private:
     llgl::Camera camera() const {
-        return llgl::Camera { llgl::Projection::perspective({ 1.44, raw_size().x() / raw_size().y(), 0.1, 20 }, Util::Recti { rect() }) }
+        return llgl::Camera {
+            llgl::Projection::perspective(
+                { 1.44, static_cast<double>(raw_size().x()) / raw_size().y(),
+                    0.1, 20 },
+                Util::Recti { rect() })
+        }
             .translate({ 0, 3, 3 })
             .rotate_x(45.0_deg)
             .translate({ 0, 0, -1 });
@@ -33,19 +39,22 @@ private:
     }
 
     virtual void draw(Gfx::Painter& painter) const override {
-        GUI::WorldDrawScope scope { painter, GUI::WorldDrawScope::ClearDepth::Yes };
+        GUI::WorldDrawScope scope { painter,
+            GUI::WorldDrawScope::ClearDepth::Yes };
 
         static Essa::Shaders::Lighting shader;
         Essa::Shaders::Lighting::Uniforms shader_uniforms;
         shader_uniforms.set_light_color(Util::Colors::LightGoldenRodYellow);
 
-        auto& model = resource_manager().require_external<Essa::Model>("../examples/essa-engine/cube.obj");
+        auto& model = resource_manager().require_external<Essa::Model>(
+            "../examples/essa-engine/cube.obj");
 
         auto transform = llgl::Transform {}
                              .rotate_x(m_angle_x.rad())
                              .rotate_y(m_angle_y.rad())
                              .rotate_z(m_angle_z.rad());
-        shader_uniforms.set_transform(transform.matrix(), camera().view_matrix(), camera().projection().matrix());
+        shader_uniforms.set_transform(transform.matrix(),
+            camera().view_matrix(), camera().projection().matrix());
         model.render(painter.renderer(), shader, shader_uniforms);
     }
 
@@ -70,7 +79,8 @@ public:
 
 private:
     virtual void update() override {
-        m_tps_container->set_content(Util::UString { "TPS: " + std::to_string(GUI::Application::the().tps()) });
+        m_tps_container->set_content(Util::UString {
+            "TPS: " + std::to_string(GUI::Application::the().tps()) });
     }
 
     GUI::Textfield* m_tps_container = nullptr;
