+++ examples/essa-gui/heighway.cpp
@@ -23,12 +23,18 @@ private:
         Gfx::RectangleDrawOptions rect;
         rect.texture = &m_framebuffer.color_texture();
         rect.texture_rect = { 0, 0, 500, 500 };
-        painter.deprecated_draw_rectangle(local_rect(), rect);
+        painter.deprecated_draw_rectangle(local_rect().cast<float>(), rect);
 
-        Gfx::Text text { Util::UString { std::to_string(m_point_count) + "; "
-                             + std::to_string(static_cast<int>(GUI::Application::the().tps())) + " FPS; "
-                             + std::to_string(std::floor(static_cast<double>(m_point_count) / (m_clock.elapsed() / 1.0s))) + " pps" },
-            resource_manager().font() };
+        Gfx::Text text {
+            Util::UString { std::to_string(m_point_count) + "; "
+                + std::to_string(
+                    static_cast<int>(GUI::Application::the().tps()))
+                + " FPS; "
+                + std::to_string(std::floor(static_cast<double>(m_point_count)
+                    / (m_clock.elapsed() / 1.0s)))
+                + " pps" },
+            resource_manager().font()
+        };
         text.set_font_size(16);
         text.set_fill_color(Util::Colors::Wheat);
         text.set_position({ 10, 20 });
@@ -43,14 +49,20 @@ private:
             m_current_position.y() = -0.4 * old_position.y() + 0.1;
         }
         else {
-            m_current_position.x() = 0.76 * old_position.x() - 0.4 * old_position.y();
-            m_current_position.y() = 0.4 * old_position.x() + 0.76 * old_position.y();
+            m_current_position.x()
+                = 0.76 * old_position.x() - 0.4 * old_position.y();
+            m_current_position.y()
+                = 0.4 * old_position.x() + 0.76 * old_position.y();
         }
 
-        m_painter.builder().set_projection(llgl::Projection { {}, { 0, 0, 500, 500 } });
+        m_painter.builder().set_projection(
+            llgl::Projection { {}, { 0, 0, 500, 500 } });
         Gfx::RectangleDrawOptions point;
         point.fill_color = Util::Colors::White;
-        m_painter.deprecated_draw_rectangle({ Util::Cs::Point2f::from_deprecated_vector(m_current_position), { 0.001, 0.001 } }, point);
+        m_painter.deprecated_draw_rectangle(
+            { Util::Cs::Point2f::from_deprecated_vector(m_current_position),
+                { 0.001, 0.001 } },
+            point);
     }
 
     llgl::Framebuffer m_framebuffer { { 500, 500 } };
@@ -62,7 +74,8 @@ private:
 
 int main() {
     GUI::Application app;
-    auto& wnd = app.create_host_window({ 750, 750 }, "Heighway's Dragon (Gfx::Painter benchmark)");
+    auto& wnd = app.create_host_window(
+        { 750, 750 }, "Heighway's Dragon (Gfx::Painter benchmark)");
 
     wnd.set_main_widget<MainWidget>();
