+++ examples/essa-gui/rich-text.cpp
@@ -16,7 +16,7 @@ private:
                 .font_size = static_cast<int>(theme().label_font_size),
                 .text_alignment = GUI::Align::Center,
             } };
-        drawable.set_rect(local_rect());
+        drawable.set_rect(local_rect().cast<float>());
         drawable.draw(painter);
     }
 
@@ -35,9 +35,11 @@ consectetur cumque non error. Harum totam, accusamus doloremque corporis rem obc
         Util::Colors::White);
 
     text.append("This thing will cost 50", Util::Colors::White);
-    text.append_image(app.resource_manager().require_texture("gui/executableFile.png"));
+    text.append_image(
+        app.resource_manager().require_texture("gui/executableFile.png"));
 
-    text.append("Lorem ipsum dolor\u00a0sit (NBSP working as expected) amet consectetur adipisicing elit. Modi \
+    text.append(
+        "Lorem ipsum dolor\u00a0sit (NBSP working as expected) amet consectetur adipisicing elit. Modi \
 iure unde optio error laboriosam exercitationem nam temporibus? Minus corporis \
 consectetur cumque non error. Harum totam, accusamus doloremque corporis rem obcaecati? ",
         Util::Colors::Red);
